Time: 3 ms (76.23%), Space: 69.6 MB (43.33%) - LeetHub

+++ 0643-maximum-average-subarray-i/0643-maximum-average-subarray-i.java
@@ -27,6 +27,6 @@ class Solution {
 
         //return double((double)max/(double)k);
 
-        return (double)((double)max/(double)k);
+        return (double)max/k;
     }
 }
